# docs/swagger.yaml
basePath: /
definitions:
  CommonResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  Link:
    properties:
      link:
        example: http://file-server.com/offers_table.xlsx
        type: string
    required:
    - link
    type: object
  common.Merchant:
    properties:
      id:
        type: integer
      offers:
        items:
          $ref: '#/definitions/common.Offer'
        type: array
    type: object
  common.Offer:
    properties:
      id:
        type: integer
      name:
        type: string
      price:
        type: integer
      quantity:
        type: integer
    type: object
  common.Task:
    properties:
      created:
        type: integer
      deleted:
        type: integer
      isCompleted:
        type: boolean
      missed:
        type: integer
      timeSpent:
        description: should be time.Duration, made as string for fancy output
        type: string
      updated:
        type: integer
    type: object
host: localhost
info:
  contact: {}
  license: {}
  title: MerchantX Trainee API
paths:
  /merchants/{id}:
    put:
      consumes:
      - application/json
      description: |-
        Проверяет полученную в теле ссылку
        Возвращает Content-Location созданной задачи по обработке
        Запускает горутину для обработки файла
      parameters:
      - description: Merchant ID
        in: path
        name: id
        required: true
        type: integer
      - description: Link to .xlsx file
        in: body
        name: merchantPutRequest
        required: true
        schema:
          $ref: '#/definitions/Link'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/CommonResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/CommonResponse'
      summary: Инициирует обработку файла
  /offers:
    get:
      parameters:
      - description: ID товара
        in: query
        name: offerId
        type: integer
      - description: ID продавца
        in: query
        name: merhcantId
        type: integer
      - description: Подстрока названия
        in: query
        name: sub
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/common.Merchant'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/CommonResponse'
      summary: Возвращает массив Продавцов с Товарами, отфильтрованными по опциональным
        параметрам
  /tasks/{id}:
    get:
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/CommonResponse'
      summary: Возвращает статистику по запущенному заданию
swagger: "2.0"